DimensionReduction/requirements.txt
python_mnist